refactor(ibc): Update dependencies and fix import errors in ics06_solomachine module
