+++ crates/ibc/Cargo.toml
@@ -35,7 +35,7 @@ mocks-no-std = ["cfg-if"]
 
 [dependencies]
 # Proto definitions for all IBC-related interfaces, e.g., connections or channels.
-ibc-proto = { default-features = false, features = ["parity-scale-codec", "borsh"], git = "https://github.com/octopus-network/ibc-proto-rs.git", rev = "e9fb075d666bf49e491e3d338e5af98aa7b2b73e"}
+ibc-proto = {version = "0.29.0", default-features = false, features = ["parity-scale-codec", "borsh"] }
 ics23 = { version = "0.9.0", default-features = false, features = ["host-functions"] }
 time = { version = ">=0.3.0, <0.3.21", default-features = false }
 serde_derive = { version = "1.0.104", default-features = false, optional = true }
@@ -61,7 +61,7 @@ scale-info = { version = "2.1.2", default-features = false, features = ["derive"
 borsh = {version = "0.10.0", default-features = false, optional = true }
 parking_lot = { version = "0.12.1", default-features = false, optional = true }
 cfg-if = { version = "1.0.0", optional = true }
-cosmrs = { version = "0.14", default-features = false, git = "https://github.com/octopus-network/cosmos-rust.git", branch = "new-proto" }
+cosmrs = { version = "0.13" }
 
 [dependencies.tendermint]
 version = "0.30"
