+++ crates/ibc/src/clients/ics06_solomachine/misbehaviour/signature_and_data.rs
@@ -1,7 +1,7 @@
 use crate::clients::ics06_solomachine::error::Error;
 use crate::clients::ics06_solomachine::types::DataType;
+use crate::core::timestamp::Timestamp;
 use crate::prelude::*;
-use crate::timestamp::Timestamp;
 use ibc_proto::ibc::lightclients::solomachine::v1::SignatureAndData as RawSignatureAndData;
 use ibc_proto::protobuf::Protobuf;
